# openapi/openapi.gen.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "e2clicker service",
    "version": "0"
  },
  "servers": [
    {
      "url": "https://e2clicker.app/api"
    },
    {
      "url": "/api"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/dosage": {
      "get": {
        "summary": "Get the user's dosage and optionally their history",
        "operationId": "dosage",
        "parameters": [
          {
            "name": "start",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time",
              "description": "The start date of the history to retrieve, if requested."
            }
          },
          {
            "name": "end",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time",
              "description": "The end date of the history to retrieve, if requested."
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved the dosage, if set.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "dosage": {
                      "description": "The user's current dosage schedule. This is null if the user has no dosage set.",
                      "allOf": [
                        {
                          "$ref": "#/components/schemas/Dosage"
                        }
                      ]
                    },
                    "history": {
                      "description": "The user's dosage history within the requested time range. If either historyStart or historyEnd are not provided, this will be null.",
                      "allOf": [
                        {
                          "$ref": "#/components/schemas/DosageHistory"
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "dosage"
        ]
      },
      "put": {
        "summary": "Set the user's dosage",
        "operationId": "setDosage",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Dosage"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successfully set the dosage."
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        },
        "tags": [
          "dosage"
        ]
      },
      "delete": {
        "summary": "Clear the user's dosage schedule",
        "operationId": "clearDosage",
        "responses": {
          "204": {
            "description": "Successfully cleared the dosage."
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        },
        "tags": [
          "dosage"
        ]
      }
    },
    "/dosage/dose": {
      "post": {
        "summary": "Record a new dosage to the user's history",
        "operationId": "recordDose",
        "description": "This endpoint is used to record a new dosage observation to the user's history. The current time is automatically used.",
        "responses": {
          "200": {
            "description": "Successfully recorded dosage.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DosageObservation"
                }
              }
            }
          }
        },
        "tags": [
          "dosage"
        ]
      },
      "put": {
        "summary": "Update a dosage in the user's history",
        "operationId": "editDose",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "description": "The updated dosage observation. The ID must match an existing observation.",
                "allOf": [
                  {
                    "$ref": "#/components/schemas/DosageObservation"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successfully updated dosage."
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        },
        "tags": [
          "dosage"
        ]
      },
      "delete": {
        "summary": "Delete multiple dosages from the user's history",
        "operationId": "forgetDoses",
        "parameters": [
          {
            "name": "dose_ids",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int64",
                "description": "The dose IDs of the observations to delete."
              }
            },
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "Successfully deleted dosages."
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        },
        "tags": [
          "dosage"
        ]
      }
    },
    "/dosage/export": {
      "get": {
        "summary": "Export the user's dosage history",
        "operationId": "exportDosageHistory",
        "parameters": [
          {
            "name": "Accept",
            "in": "header",
            "schema": {
              "type": "string",
              "enum": [
                "text/csv",
                "application/json"
              ],
              "default": "text/csv"
            },
            "required": true,
            "description": "The format to export the dosage history in."
          },
          {
            "name": "start",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time",
              "description": "The start date of the history to retrieve. If not provided, defaults to the beginning of time."
            }
          },
          {
            "name": "end",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time",
              "description": "The end date of the history to retrieve. If not provided, defaults to the current time."
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved the dosage history.",
            "content": {
              "text/csv": {
                "schema": {
                  "$ref": "#/components/schemas/DosageHistoryCSV"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DosageHistory"
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimitedResponse"
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        },
        "tags": [
          "dosage"
        ]
      }
    },
    "/dosage/import": {
      "post": {
        "summary": "Import a CSV file of dosage history",
        "operationId": "importDosageHistory",
        "parameters": [
          {
            "name": "Content-Type",
            "in": "header",
            "schema": {
              "type": "string",
              "enum": [
                "text/csv",
                "application/json"
              ],
              "default": "text/csv"
            },
            "required": true,
            "description": "The format to import the dosage history as."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "text/csv": {
              "schema": {
                "$ref": "#/components/schemas/DosageHistoryCSV"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DosageHistory"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully imported the dosage history.",
            "content": {
              "application/json": {
                "schema": {
                  "required": [
                    "records",
                    "succeeded"
                  ],
                  "properties": {
                    "records": {
                      "type": "integer",
                      "description": "The number of records in the file."
                    },
                    "succeeded": {
                      "type": "integer",
                      "description": "The number of records actually imported successfully. This is not equal to #records if there were errors or duplicate entries."
                    },
                    "error": {
                      "$ref": "#/components/schemas/Error"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "dosage"
        ]
      }
    },
    "/pushinfo": {
      "get": {
        "summary": "Get the server's push notification information",
        "operationId": "webPushInfo",
        "security": [],
        "responses": {
          "200": {
            "description": "Successfully retrieved the server's push notification information.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PushInfo"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        },
        "tags": [
          "notification"
        ]
      }
    },
    "/notifications/methods": {
      "get": {
        "summary": "Get the user's notification methods",
        "operationId": "userNotificationMethods",
        "responses": {
          "200": {
            "description": "Successfully retrieved the user's push notification subscription.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReturnedNotificationMethods"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        },
        "tags": [
          "notification"
        ]
      }
    },
    "/notifications/methods/push/{deviceID}": {
      "get": {
        "summary": "Get the user's push notification subscription",
        "operationId": "userPushSubscription",
        "parameters": [
          {
            "name": "deviceID",
            "in": "path",
            "schema": {
              "$ref": "#/components/schemas/PushDeviceID"
            },
            "required": true,
            "description": "The device ID of the push subscription to retrieve."
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved the user's push notification subscription. The returned object will contain secrets.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PushSubscription"
                }
              }
            }
          },
          "404": {
            "description": "The user does not have a push notification subscription for the specified device ID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        },
        "tags": [
          "notification"
        ]
      },
      "delete": {
        "summary": "Unsubscribe from push notifications",
        "operationId": "userUnsubscribePush",
        "parameters": [
          {
            "name": "deviceID",
            "in": "path",
            "schema": {
              "$ref": "#/components/schemas/PushDeviceID"
            },
            "required": true,
            "description": "The device ID of the push subscription to unsubscribe from."
          }
        ],
        "responses": {
          "204": {
            "description": "Successfully unsubscribed from push notifications."
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        },
        "tags": [
          "notification"
        ]
      }
    },
    "/notifications/methods/push": {
      "put": {
        "summary": "Create or update a push subscription",
        "operationId": "userSubscribePush",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PushSubscription"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successfully subscribed to push notifications."
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        },
        "tags": [
          "notification"
        ]
      }
    },
...
    "/me": {
      "get": {
        "summary": "Get the current user",
        "operationId": "currentUser",
        "responses": {
          "200": {
            "description": "Successfully retrieved the current user.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/User"
                    },
                    {
                      "type": "object",
                      "required": [
                        "secret"
                      ],
                      "properties": {
                        "secret": {
                          "$ref": "#/components/schemas/UserSecret"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        },
        "tags": [
          "user"
        ]
      }
    },
    "/me/avatar": {
      "get": {
        "summary": "Get the current user's avatar",
        "operationId": "currentUserAvatar",
        "responses": {
          "200": {
            "description": "Successfully retrieved the user's avatar.",
            "content": {
              "image/*": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        },
        "tags": [
          "user"
        ]
      },
      "put": {
        "summary": "Set the current user's avatar",
        "operationId": "setCurrentUserAvatar",
        "requestBody": {
          "required": true,
          "content": {
            "image/*": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successfully set the user's avatar."
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        },
        "tags": [
          "user"
        ]
      }
    },
    "/me/sessions": {
      "get": {
        "summary": "List the current user's sessions",
        "operationId": "currentUserSessions",
        "responses": {
          "200": {
            "description": "Successfully retrieved the user's sessions.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Session"
                  }
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        },
        "tags": [
          "user"
        ]
      },
      "delete": {
        "summary": "Delete one of the current user's sessions",
        "operationId": "deleteUserSession",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "description": "The session identifier to delete"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successfully deleted the user's sessions."
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        },
        "tags": [
          "user"
        ]
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "Dosage": {
        "type": "object",
        "required": [
          "deliveryMethod",
          "dose",
          "interval"
        ],
        "properties": {
          "deliveryMethod": {
            "type": "string",
            "description": "The delivery method to use.",
            "x-order": 1
          },
          "dose": {
            "type": "number",
            "format": "float",
            "description": "The dosage amount.",
            "x-order": 2
          },
          "interval": {
            "type": "number",
            "format": "double",
            "description": "The interval between doses in days.",
            "x-order": 3
          },
          "concurrence": {
            "type": "integer",
            "description": "The number of estrogen patches on the body at once. Only relevant if delivery method is patch.",
            "x-order": 4
          }
        }
      },
      "DosageHistory": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/DosageObservation"
        }
      },
      "DosageHistoryCSV": {
        "type": "string",
        "example": "dose,method,takenAt,takenOffAt,comment\n100,patch tw,2020-01-01T00:00:00Z,,\n100,patch tw,2020-01-02T12:00:00Z,,\n100,patch tw,2020-01-04T00:00:00Z,,\n100,patch tw,2020-01-05T12:00:00Z,,",
        "description": "The CSV format of the user's dosage history."
      },
      "DosageObservation": {
        "type": "object",
        "required": [
          "id",
          "deliveryMethod",
          "dose",
          "takenAt"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "The unique identifier for the observation.",
            "x-order": 1
          },
          "deliveryMethod": {
            "type": "string",
            "description": "The delivery method used.",
            "x-order": 2
          },
          "dose": {
            "type": "number",
            "format": "float",
            "description": "The dosage amount.",
            "x-order": 3
          },
          "takenAt": {
            "type": "string",
            "format": "date-time",
            "description": "The time the dosage was taken.",
            "x-order": 4
          },
          "takenOffAt": {
            "type": "string",
            "format": "date-time",
            "description": "The time the dosage was taken off. This is only relevant for patch delivery methods.",
            "x-order": 5
          },
          "comment": {
            "type": "string",
            "description": "A comment about the dosage, if any.",
            "x-order": 6
          }
        }
      },
...
      "ReturnedNotificationMethods": {
        "properties": {
          "webPush": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReturnedPushSubscription"
            }
          }
        }
      },
...
      "UserSecret": {
        "description": "A secret and unique user identifier. This secret is generated once and never changes. It is used to both authenticate and identify a user, so it should be kept secret.",
        "type": "string",
        "x-go-type": "user.Secret",
        "x-go-type-import": {
          "path": "libdb.so/e2clicker/services/user",
          "name": "userservice"
        }
      },
...
    },
...
    "responses": {
      "ErrorResponse": {
        "description": "The request is invalid.\n",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "RateLimitedResponse": {
        "description": "The request has been rate limited.\n",
        "headers": {
          "Retry-After": {
            "$ref": "#/components/headers/Retry-After"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  }
}